% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfda_utils.R
\name{bin_latent_variable}
\alias{bin_latent_variable}
\title{Bins latent variable data}
\usage{
bin_latent_variable(latent_variable_vector, depth_vector = c(0, 1, 2, 5, 10))
}
\arguments{
\item{latent_variable_vector}{Vector of (not-yet-binned) latent variable values}

\item{depth_vector}{Depths to consider (e.g the depths we are using in inferrence code)}
}
\value{
A numeric vector of binned data for the latent variable
}
\description{
This is needed since our inferred data is binned, so for power analyses
we would also like to have less resolution.
}
