% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_utils.R
\name{draws_from_prior}
\alias{draws_from_prior}
\title{Draws from list that contains dist and params}
\usage{
draws_from_prior(prior_sub_list, num_samples = 10000)
}
\arguments{
\item{prior_sub_list, }{contains mapping both a list of dist and params
e.g. list("dist" = "normal", "params" = c(0,1))}

\item{num_samples, }{optional number of samples to draw}
}
\value{
Samples drawn from distributoin
}
\description{
Draws from list that contains dist and params
}
