% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_utils.R
\name{get_inclusion_bayes_factors_for_models}
\alias{get_inclusion_bayes_factors_for_models}
\title{Calculates Inclusion Bayes Factors given Formulas, Data}
\usage{
get_inclusion_bayes_factors_for_models(fitted_models, match_models = FALSE)
}
\arguments{
\item{fitted_models}{Fitted models}

\item{match_models}{see bayestestR documentation}
}
\value{
A list with model Bayes factors as 'model_bfs' and
inclusion Bayes factors as 'inclusion_bfs'
}
\description{
This function calculates Inclusion Bayes Factors for each variable.
}
