% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_utils.R
\name{get_text_prior}
\alias{get_text_prior}
\title{Loads in prior list}
\usage{
get_text_prior(prior_sub_list)
}
\arguments{
\item{prior_sub_list, }{contains mapping both a list of dist and params
e.g. list("dist" = "normal", "params" = c(0,1))}
}
\value{
Text prior for brms
}
\description{
Loads in prior list
}
