% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_utils.R
\name{plot_dist}
\alias{plot_dist}
\title{Plots a distribution to comply with WAMBS guidelines}
\usage{
plot_dist(prior_sub_list, num_samples = 10000, save_path = "./test.png")
}
\arguments{
\item{prior_sub_list, }{contains mapping both a list of dist and params
e.g. list("dist" = "normal", "params" = c(0,1))}

\item{num_samples, }{optional number of samples to draw}

\item{save_path, }{where to save png}
}
\value{
Nothing, saves plot to save_path
}
\description{
Plots a distribution to comply with WAMBS guidelines
}
