% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_utils.R
\name{plotting_dist_dict}
\alias{plotting_dist_dict}
\title{Provides mapping of brms/stan -> plotting distributions}
\usage{
plotting_dist_dict()
}
\value{
list mapping brms/stan distribution name to a distribution
for plotting or drawing from (as text)
}
\description{
This could be added as data, but for now it is added as a function.
To extend functionality, add more distributions in the list.
}
